House Cleaning, ready for last stage.
